+++ csaps/_base.py
@@ -10,7 +10,7 @@ import typing as ty
 
 import numpy as np
 
-from csaps._types import TData, TProps, TSmooth, TXi, TSpline
+from ._types import TData, TProps, TSmooth, TXi, TSpline
 
 
 class SplinePPFormBase(abc.ABC, ty.Generic[TData, TProps]):
@@ -87,17 +87,15 @@ class SplinePPFormBase(abc.ABC, ty.Generic[TData, TProps]):
             Interpolated/smoothed data
         """
 
-    def __repr__(self):
+    def __repr__(self):  # pragma: no cover
         return (
-            '{}\n'
-            '  breaks: {}\n'
-            '  coeffs: {} shape\n'
-            '  pieces: {}\n'
-            '  order: {}\n'
-            '  ndim: {}\n'
-        ).format(
-            type(self).__name__, self.breaks, self.coeffs.shape,
-            self.pieces, self.order, self.ndim)
+            f'{type(self).__name__}\n'
+            f'  breaks: {self.breaks}\n'
+            f'  coeffs: {self.coeffs.shape} shape\n'
+            f'  pieces: {self.pieces}\n'
+            f'  order: {self.order}\n'
+            f'  ndim: {self.ndim}\n'
+        )
 
 
 class ISmoothingSpline(abc.ABC, ty.Generic[TSpline, TSmooth, TXi]):